requirements.txt
openai>=1.1.0
yfinance>=0.2.12
pandas>=1.3.5
numpy>=1.20.0
requests>=2.27.1
python-dotenv>=0.20.0
cachetools>=5.0.0
ratelimit>=2.2.1
backoff>=2.1.2
ta-lib>=0.4.0; platform_system!="Windows"